requirements.txt
fastmcp>=2.0.0
uv>=0.1.0
fastapi>=0.100.0
uvicorn>=0.24.0
elasticsearch
python-dotenv